+++ examples/example-no-command.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env -S deno run -A
-import { cliteRun } from "jsr:@jersou/clite@0.5.0";
+import { cliteRun } from "jsr:@jersou/clite@0.6.0";
 import { Tool } from "./example.ts";
 
 if (import.meta.main) { // if the file is imported, do not execute this block